Prepared bootstrapping for v5.4
