+++ pimcore/config/kernel.php
@@ -11,19 +11,6 @@
  * @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
-use Pimcore\Config;
-use Symfony\Component\Debug\Debug;
 
-$environment = Config::getEnvironment();
-$debug       = Config::getEnvironmentConfig()->activatesKernelDebugMode($environment);
-
-if ($debug) {
-    Debug::enable();
-    @ini_set('display_errors', 'On');
-}
-
-$kernel = new AppKernel($environment, $debug);
-Pimcore::setKernel($kernel);
-$kernel->boot();
-
+$kernel = \Pimcore\Bootstrap::kernel();
 return $kernel;
