+++ pimcore/config/startup.php
@@ -12,21 +12,5 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-// this should already be defined at this point, but we include a fallback here
-// fot backwards compatibility
-if (!defined('PIMCORE_PROJECT_ROOT')) {
-    define(
-        'PIMCORE_PROJECT_ROOT',
-        getenv('PIMCORE_PROJECT_ROOT')
-            ?: getenv('REDIRECT_PIMCORE_PROJECT_ROOT')
-            ?: realpath(__DIR__ . '/../..')
-    );
-}
-
-// set up constants and autoloading
-require_once  __DIR__ . '/bootstrap.php';
-
-// load kernel
-$kernel = require_once  __DIR__ . '/kernel.php';
-
+$kernel = \Pimcore\Bootstrap::startup();
 return $kernel;
